refactor(server): refactor the project struct of server
