+++ server/src/app.module.ts
@@ -1,10 +1,7 @@
 import { Module } from '@nestjs/common'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
-import { CollectionsModule } from './collections/collections.module'
 import { WebsitesModule } from './websites/websites.module'
-import { BucketsModule } from './buckets/buckets.module'
-import { PoliciesModule } from './policies/policies.module'
 import { FunctionsModule } from './functions/functions.module'
 import { HttpModule } from '@nestjs/axios'
 import { CoreModule } from './core/core.module'
@@ -14,6 +11,9 @@ import { ThrottlerModule } from '@nestjs/throttler'
 import { InitializerModule } from './initializer/initializer.module'
 import { InstanceModule } from './instance/instance.module'
 import { ScheduleModule } from '@nestjs/schedule'
+import { DatabaseModule } from './database/database.module'
+import { PrismaService } from './prisma.service'
+import { StorageModule } from './storage/storage.module'
 
 @Module({
   imports: [
@@ -23,18 +23,17 @@ import { ScheduleModule } from '@nestjs/schedule'
       limit: 10,
     }),
     FunctionsModule,
-    PoliciesModule,
-    BucketsModule,
     WebsitesModule,
-    CollectionsModule,
     HttpModule,
     AuthModule,
     CoreModule,
     ApplicationsModule,
     InitializerModule,
     InstanceModule,
+    DatabaseModule,
+    StorageModule,
   ],
   controllers: [AppController],
-  providers: [AppService],
+  providers: [AppService, PrismaService],
 })
 export class AppModule {}
