+++ server/src/applications/applications.service.ts
@@ -9,10 +9,8 @@ import { DatabaseCoreService } from 'src/core/database.cr.service'
 import { GatewayCoreService } from 'src/core/gateway.cr.service'
 import { OSSUserCoreService } from 'src/core/oss-user.cr.service'
 import { APPLICATION_SECRET_KEY, ServerConfig } from 'src/constants'
-import { GenerateAlphaNumericPassword } from 'src/common/random'
+import { GenerateAlphaNumericPassword } from 'src/utils/random'
 import { OSSUser } from 'src/core/api/oss-user.cr'
-import * as assert from 'node:assert'
-import { JwtService } from '@nestjs/jwt'
 
 @Injectable()
 export class ApplicationsService {
@@ -22,7 +20,6 @@ export class ApplicationsService {
     private readonly databaseCore: DatabaseCoreService,
     private readonly gatewayCore: GatewayCoreService,
     private readonly ossCore: OSSUserCoreService,
-    private readonly jwtService: JwtService,
   ) {}
 
   async create(userid: string, dto: CreateApplicationDto) {
@@ -211,26 +208,5 @@ export class ApplicationsService {
       ],
     }
     return JSON.stringify(policy)
-  }
-
-  async getDebugFunctionToken(appid: string) {
-    const conf = await this.prisma.applicationConfiguration.findUnique({
-      where: { appid },
-    })
-
-    // get secret from envs
-    const secret = conf?.environments.find(
-      (env) => env.name === APPLICATION_SECRET_KEY,
-    )
-    assert(secret?.value, 'application secret not found')
-
-    // generate token
-    const exp = Math.floor(Date.now() / 1000) + 60 * 60 * 24 * 7 // 7 days
-
-    const token = this.jwtService.sign(
-      { appid, type: 'debug', exp },
-      { secret: secret.value },
-    )
-    return token
   }
 }
