+++ server/src/core/oss-user.cr.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, Logger } from '@nestjs/common'
-import { GetApplicationNamespaceById } from '../common/getter'
+import { GetApplicationNamespaceById } from '../utils/getter'
 import { KubernetesService } from './kubernetes.service'
 import * as assert from 'node:assert'
 import { Bundle, Application } from '@prisma/client'
 import { MB, ResourceLabels } from 'src/constants'
-import { GenerateAlphaNumericPassword } from 'src/common/random'
-import { toQuantityString } from 'src/common/types'
+import { GenerateAlphaNumericPassword } from 'src/utils/random'
+import { toQuantityString } from 'src/utils/types'
 import { OSSUser } from './api/oss-user.cr'
 
 @Injectable()
