+++ server/src/database/collections.controller.ts
@@ -17,7 +17,7 @@ import {
 } from '@nestjs/swagger'
 import { ApplicationAuthGuard } from '../auth/application.auth.guard'
 import { JwtAuthGuard } from '../auth/jwt.auth.guard'
-import { ApiResponseUtil, ResponseUtil } from '../common/response'
+import { ApiResponseUtil, ResponseUtil } from '../utils/response'
 import { CollectionsService } from './collections.service'
 import { CreateCollectionDto } from './dto/create-collection.dto'
 import { UpdateCollectionDto } from './dto/update-collection.dto'
