+++ server/src/database/dto/update-collection.dto.ts
