+++ server/src/functions/functions.module.ts
@@ -1,4 +1,5 @@
 import { Module } from '@nestjs/common'
+import { JwtService } from '@nestjs/jwt'
 import { ApplicationsModule } from 'src/applications/applications.module'
 import { PrismaService } from 'src/prisma.service'
 import { CoreModule } from '../core/core.module'
@@ -8,6 +9,6 @@ import { FunctionsService } from './functions.service'
 @Module({
   imports: [CoreModule, ApplicationsModule],
   controllers: [FunctionsController],
-  providers: [FunctionsService, PrismaService],
+  providers: [FunctionsService, PrismaService, JwtService],
 })
 export class FunctionsModule {}
