--- server/src/policies/policies.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common'
-import { PoliciesService } from './policies.service'
-import { PoliciesController } from './policies.controller'
-
-@Module({
-  controllers: [PoliciesController],
-  providers: [PoliciesService],
-})
-export class PoliciesModule {}
