+++ server/src/storage/storage.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common'
-import { BucketsController } from './buckets.controller'
+import { BucketsController } from '../storage/buckets.controller'
 import { CoreModule } from '../core/core.module'
-import { ApplicationsModule } from 'src/applications/applications.module'
+import { ApplicationsModule } from '../applications/applications.module'
 
 @Module({
   imports: [CoreModule, ApplicationsModule],
   controllers: [BucketsController],
   providers: [],
 })
-export class BucketsModule {}
+export class StorageModule {}
