+++ server/src/utils/random.ts
